Website/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
SECRET_KEY=your-secret-key-here 